- Fix DDEServer to work again! - Use proper encoding - Click area only on focused vessels

+++ Src/Orbiter/ddeserver.h
@@ -45,7 +45,7 @@ protected:
 	// Send a vessel list to hClient
 	void VesselList (HWND hClient);
 
-	DDEDATA *AllocDatablock (DWORD bufsize);
+	DDEDATA *AllocDatablock (size_t bufsize);
 
 	bool isVessel (Vessel *v);
 	bool isBody (Body *b);
